Update devscripts.sh for homebrew

+++ devscripts.sh
@@ -1,18 +1,22 @@
 #!/bin/bash
 PYTHON="/usr/bin/env python3"
-EXEC="zoomdl"
-# ZIP="zoomdl.zip"
-ZIP_FOLDER="zip"
 LOCALBIN="/usr/local/bin/"
-SRC="zoom_dl"
+EXEC="zoomdl"
+EXEC_FOLDER="build"
+ZIP_FOLDER="zip"
+SRC_FOLDER="zoomdl"
+HOMEBREW_FLAG=false
 
 compile(){
-	mkdir -p $ZIP_FOLDER #create folder to zip
-	cp -r $SRC $ZIP_FOLDER #copy all python files
-	mv $ZIP_FOLDER/$SRC/__main__.py $ZIP_FOLDER #put main above src
-	pip install -r requirements.txt --target $ZIP_FOLDER
-	rm -r $ZIP_FOLDER/*.dist-info
-	python3 -m  zipapp -p "$PYTHON" -c -o $EXEC $ZIP_FOLDER
+	mkdir -p $EXEC_FOLDER  # create folder for executable
+	mkdir -p $ZIP_FOLDER  #create folder to zip
+	cp -r $SRC_FOLDER $ZIP_FOLDER  #copy all python files
+	mv $ZIP_FOLDER/$SRC_FOLDER/__main__.py $ZIP_FOLDER  #put main above src
+	if [ "$HOMEBREW_FLAG" = false ]; then
+		pip install -r requirements.txt --target $ZIP_FOLDER
+		rm -r $ZIP_FOLDER/*.dist-info
+	fi
+	python3 -m zipapp -p "$PYTHON" -c -o $EXEC_FOLDER/$EXEC $ZIP_FOLDER
 	rm -rf $ZIP_FOLDER
 }
 
@@ -20,9 +24,14 @@ build(){
 	compile
 }
 
+homebrew(){
+	HOMEBREW_FLAG=true
+	compile
+}
+
 install(){
 	compile
-	sudo cp $EXEC $LOCALBIN
+	sudo cp $EXEC_FOLDER/$EXEC $LOCALBIN
 }
 
 if [[ $# -ne 1 ]]
